chore: extract methods for agent enroll to its own file
